Fix error with SubjectCreate

+++ src/main/java/pl/byrka/uczelnia/model/Entity/Subject/SubjectCreate.java
@@ -12,7 +12,7 @@ public class SubjectCreate {
     public String name;
     public int ects;
     public typeEnum type;
-    public int lecturer;
+    public long lecturer;
 
     public enum typeEnum{
         Wyklad("Wykład"),
